# openapi.yaml
openapi: 3.0.3
info:
  title: Expense Approval API
  version: 1.0.0
  description: Contract used as the ground truth for integration tests.
paths:
  /expenses:
    get:
      summary: List all expenses (newest first)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Expense' }
    post:
      summary: Submit an expense (R1/R2/R3)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '400': { $ref: '#/components/responses/Error' }
  /expenses/{id}/approve:
    post:
      summary: Approve a Pending expense (R4/R5/R6)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApproveInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /expenses/{id}/reject:
    post:
      summary: Reject a Pending expense with a reason (R7)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
    Status:
      type: string
      enum: [Pending, Approved, Rejected]
    SubmitInput:
      type: object
      required: [amount, category, submitterId]
      properties:
        amount: { type: number }
        category: { type: string, enum: [Travel, Meals, Equipment, Other] }
        submitterId: { type: string }
    ApproveInput:
      type: object
      required: [approverId]
      properties:
        approverId: { type: string }
        approverRole: { type: string, enum: [Employee, Manager] }
    RejectInput:
      type: object
      required: [approverId, reason]
      properties:
        approverId: { type: string }
        approverRole: { type: string, enum: [Employee, Manager] }
        reason: { type: string }
    Expense:
      type: object
      properties:
        id: { type: integer }
        amount: { type: number }
        category: { type: string }
        submitterId: { type: string }
        status: { $ref: '#/components/schemas/Status' }
        reason: { type: string, nullable: true }
        approverId: { type: string, nullable: true }